# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym ",
    "description": "JSON REST API for Dynamic Gym website "
  },
  "servers": [
    {
      "url": "http://localhost:8090"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/booking": {
      "post": {
        "summary": "Create a booking",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new booking",
          "content": {
            "application/json": {
              "schema": {
                "booking_user_id": "number",
                "booking_class_id": "number",
                "booking_created_date_time": "string"
              },
              "example": {
                "booking_user_id": 1,
                "booking_class_id": 1,
                "booking_created_date_time": "2023/10/20"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all booking",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Update an existing booking",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update an existing booking",
          "content": {
            "application/json": {
              "schema": {
                "booking_user_id": "number",
                "booking_class_id": "number",
                "booking_created_date_time": "string",
                "booking_id": "number"
              },
              "example": {
                "booking_user_id": 110,
                "booking_class_id": 110,
                "booking_created_date_time": "2023/04/10",
                "booking_id": 1
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific booking by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deleting a booking",
          "content": {
            "application/json": {
              "schema": {
                "booking_id": "number"
              },
              "example": {
                "booking_id": "1"
              }
            }
          }
        }
      }
    },
    "/booking/{id}": {
      "get": {
        "summary": "Get a specific booking by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/booking/class/{id}": {
      "get": {
        "summary": "Get a specific booking by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/booking/staff/{id}": {
      "get": {
        "summary": "Get a specific booking by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}